feat. Specialize UUIDUtils per platform
