+++ core/js/src/main/scala/tech/ant8e/uuid4cats/PlatformSpecificMethods.scala
