+++ core/jvm/src/test/scala/tech/ant8e/uuid4cats/ExtractorSuite.scala
@@ -1,0 +1,27 @@
+/*
+ * Copyright 2023 Antoine Comte
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package tech.ant8e.uuid4cats
+import java.time.Instant
+import java.util.UUID
+
+class ExtractorSuite extends ExtractorSuiteBase[Instant] {
+  override def extractTimestamp(uuid: UUID): Option[Instant] =
+    UUIDUtil.extractTimestamp(uuid)
+
+  override def timeStampFromEpochMillis(epochMillis: Long): Instant =
+    Instant.ofEpochMilli(epochMillis)
+}
