+++ core/shared/src/main/scala/tech/ant8e/uuid4cats/UUIDBuilder.scala
