+++ core/shared/src/test/scala/tech/ant8e/uuid4cats/BuilderSuite.scala
